Move `editor/` files
