+++ packages/devtools_app/lib/src/framework/theme_manager.dart
@@ -9,7 +9,7 @@ import 'package:devtools_app_shared/utils.dart';
 import 'package:dtd/dtd.dart';
 import 'package:logging/logging.dart';
 
-import '../service/editor/api_classes.dart';
+import '../shared/editor/api_classes.dart';
 import '../shared/config_specific/post_message/post_message.dart';
 import '../shared/globals.dart';
