+++ packages/devtools_app/lib/src/standalone_ui/vs_code/devtools/extensions_view.dart
@@ -9,8 +9,8 @@ import 'package:devtools_shared/devtools_extensions.dart';
 import 'package:flutter/material.dart';
 
 import '../../../extensions/extension_screen.dart';
-import '../../../service/editor/api_classes.dart';
-import '../../../service/editor/editor_client.dart';
+import '../../../shared/editor/api_classes.dart';
+import '../../../shared/editor/editor_client.dart';
 import '../../../shared/analytics/analytics.dart' as ga;
 import '../../../shared/analytics/constants.dart' as gac;
 import 'shared.dart';
