+++ packages/devtools_app/lib/src/standalone_ui/vs_code/devtools/sidebar_extensions_controller.dart
@@ -11,7 +11,7 @@ import 'package:logging/logging.dart';
 
 import '../../../extensions/extension_service.dart';
 import '../../../extensions/extension_service_helpers.dart';
-import '../../../service/editor/api_classes.dart';
+import '../../../shared/editor/api_classes.dart';
 import '../../../shared/globals.dart';
 
 final _devToolsExtensionsLog = Logger('Flutter Sidebar - DevTools Extensions');
