+++ packages/devtools_app/test/standalone_ui/vs_code/devtools_test.dart
@@ -3,7 +3,7 @@
 // found in the LICENSE file.
 
 import 'package:devtools_app/devtools_app.dart';
-import 'package:devtools_app/src/service/editor/api_classes.dart';
+import 'package:devtools_app/src/shared/editor/api_classes.dart';
 import 'package:devtools_app/src/shared/development_helpers.dart';
 import 'package:devtools_app/src/standalone_ui/vs_code/devtools/devtools_view.dart';
 import 'package:devtools_app_shared/service.dart';
